feat: search users by name in the wallet section
### Issue 6

Before this implementation, to transfer funds to another user, the current logged-in user had to enter the database ID of the target user, which wasn't good practice.

### Solution

With this pull, a user is now able to search for a given user by username/email and confirm the target user before sending the funds to them.